Fix global_histogram, move auxiliary functions into a separate file
